# src/main/java/it/gov/pagopa/apiconfig/cache/model/AppInfo.java
package it.gov.pagopa.apiconfig.cache.model;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.ToString;

import javax.validation.constraints.NotNull;

@Data
@Builder(toBuilder = true)
@NoArgsConstructor
@AllArgsConstructor(access = AccessLevel.PRIVATE)
@ToString
@JsonIgnoreProperties(ignoreUnknown = true)
public class AppInfo {

  @NotNull private String name;
  @NotNull private String version;
  @NotNull private String environment;
  private String dbConnection;
}